+++ openbank-engagement-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.2.1](https://github.com/JiRaska/open-bank-oss/compare/engagement-service-v0.2.0...engagement-service-v0.2.1) (2026-08-07)
+
+
+### Security
+
+* **engagement:** add edge-service-engagement authz policy ([#4054](https://github.com/JiRaska/open-bank-oss/issues/4054)) ([6b3594b](https://github.com/JiRaska/open-bank-oss/commit/6b3594b2b54cd4c2522f748daec593173ab27664))
+
 ## [0.2.0](https://github.com/JiRaska/open-bank-oss/compare/engagement-service-v0.1.0...engagement-service-v0.2.0) (2026-08-07)